linux-user: ignore signals 33 and 64 to allow golang emulation
Signal 33 will always fail. This causes golang crash since
https://github.com/golang/go/commit/675eb72c285cd0dd44a5f280bb3fa456ddf6de16

As explained in that commit, these signals are very rarely used in a
way that causes problems, so it's ok-ish to ignore one of them.

Signal 64 will fail because QEMU uses SIGRTMAX for itself. This causes
golang to crash for versions earlier than
https://github.com/golang/go/commit/d10675089d74db0408f2432eae3bd89a8e1c2d6a

Since after that commit golang ignores that signal, we also ignore it here to
allow earlier versions to run as well.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>
Origin: https://git.alpinelinux.org/aports/plain/main/qemu/ignore-signals-33-and-64-to-allow-golang-emulation.patch